+++ lib/pocket/client.rb
@@ -4,10 +4,31 @@ module Pocket
   # @note All methods have been separated into modules and follow the same grouping used in {TODO:doc_URL the Pocket API Documentation}.
   # @see TODO:doc_url
   class Client < API
-    Dir[File.expand_path("../client/*.rb", __FILE__)].sort.each { |f| require f }
+    # http://getpocket.com/developer/docs/v3/add
+    module Add
+      # required params: url, consumer_key, access_token
+      def add params
+        response = connection.post("/v3/add", params)
+        response.body
+      end
+    end
 
-    include Pocket::Client::Add
-    include Pocket::Client::Modify
-    include Pocket::Client::Retrieve
+    # http://getpocket.com/developer/docs/v3/modify
+    module Modify
+      # required params: actions, consumer_key, access_token
+      def modify actions
+        response = connection.post("/v3/send", {actions: actions})
+        response.body
+      end
+    end
+
+    # http://getpocket.com/developer/docs/v3/retrieve
+    module Retrieve
+      # required params: consumer_key, access_token
+      def retrieve params = {}
+        response = connection.post("/v3/get", params)
+        response.body
+      end
+    end
   end
 end
